some fix

+++ ultimatepp/UppVector.hpp
@@ -78,12 +78,12 @@ struct CollectionAdapter< Upp::Vector<T> > {
 		return 0;
 	}
 
-    static ValueType& getByIndex(ThisCol& c, long i) {
-        return c[i];
+    static ValueType& getByIndex(ThisCol& c,long i) {
+        return c[static_cast<int>(i)];
     }
 
-    static ConstValueType& getByIndex(ConstCol& c, long i) {
-        return c[i];
+    static ConstValueType& getByIndex(ConstCol& c,long i) {
+        return c[static_cast<int>(i)];
     }
 
     static void erase(ThisCol& c,Iterator pos) {
@@ -182,7 +182,7 @@ struct CollectionAdapter< const Upp::Vector<T> > {
 	}
 
     static ValueType& getByIndex(ThisCol& c, long i) {
-        return c[i];
+        return c[static_cast<int>(i)];
     }
 
     static Iterator iterAt(ConstCol& c,long pos) {
